Extract prepared texture asset planning services
