+++ tests/PlateauResoniteLink.Tests/Targets/ResoniteLiveSceneImportTargetLifecycleTests.cs
@@ -40,8 +40,10 @@ public sealed class ResoniteLiveSceneImportTargetLifecycleTests
         IResoniteMaterialPlanning materialPlanning)
     {
         return new ResonitePreparedCityObjectImporter(
-            new ResoniteGeometryAssetPlanner(new ResoniteGeometryAssetAssembler()),
-            new ResoniteSceneMaterialPlanComposer(materialPlanning),
+            new ResonitePreparedCityObjectAssetPlanner(
+                new ResonitePreparedTextureUploader(new ResoniteSharedTerrainTextureAssetWriter()),
+                new ResoniteGeometryAssetPlanner(new ResoniteGeometryAssetAssembler()),
+                new ResoniteSceneMaterialPlanComposer(materialPlanning)),
             new ResoniteBatchEmissionPlanner(),
             new PlannedBatchEmissionInterpreter());
     }
